MindHaven/requirements.txt
Django>=3.2,<4.0
pymongo[srv]>=4.0.0,<5.0.0
dnspython>=2.1.0,<3.0.0
djangorestframework>=3.12.0,<3.13.0
python-dotenv>=0.19.0,<1.0.0 